Remove intermediary file
